pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "livekit-wakeword"
dynamic = ["version"]
description = "Wake word detection for voice-enabled applications"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "Binh Pham", email = "binh.pham@livekit.io" }]
keywords = ["wake-word", "keyword-spotting", "voice", "speech", "livekit", "onnx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24",
    "onnxruntime>=1.17",
]

[project.optional-dependencies]
listener = [
    "pyaudio>=0.2.14",
]
train = [
    "torch>=2.5",
    "torchaudio>=2.5",
    "pydantic>=2.0",
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "huggingface-hub>=0.20",
    "soundfile>=0.12",
    "torch-audiomentations>=0.11",
    "audiomentations>=0.36",
    "webrtcvad>=2.0.10",
    "nltk>=3.8",
    "pronouncing>=0.2",
    "scipy>=1.11",
    "setuptools>=69.0,<81",
    "librosa>=0.10",
    "scikit-learn>=1.3",
    "cmudict>=1.0",
    "tqdm>=4.60",
]
voxcpm = [
    "voxcpm>=2.0.0",
]
eval = [
    "matplotlib>=3.8",
]
export = [
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "onnxscript>=0.6.2",
]
[project.urls]
Homepage = "https://github.com/livekit/livekit-wakeword"
Repository = "https://github.com/livekit/livekit-wakeword"
Issues = "https://github.com/livekit/livekit-wakeword/issues"

[project.scripts]
livekit-wakeword = "livekit.wakeword.cli:app"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/livekit/wakeword/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/livekit"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

# Include bundled ONNX models for feature extraction
[tool.hatch.build.targets.wheel.force-include]
"src/livekit/wakeword/resources/melspectrogram.onnx" = "livekit/wakeword/resources/melspectrogram.onnx"
"src/livekit/wakeword/resources/embedding_model.onnx" = "livekit/wakeword/resources/embedding_model.onnx"

[tool.hatch.build.targets.sdist]
include = [
    "src/livekit/wakeword/**/*.py",
    "src/livekit/wakeword/resources/*.onnx",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=4.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

# Vendored VITS graph (rhasspy/piper); keep upstream naming.
[tool.ruff.lint.per-file-ignores]
"src/livekit/wakeword/data/piper/vits/**" = ["E741"]

[tool.mypy]
python_version = "3.11"
strict = true
